sprints/sprint-2-agents/single_tool_agents/langchain/requirements.txt
langchain
openai
requests
beautifulsoup4
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
